TFT-Companion/requirements.txt
anthropic>=0.40.0
mss>=9.0.1
keyboard>=0.13.5
python-dotenv>=1.0.0
pyyaml>=6.0
pillow>=10.0.0

# OCR / vision pipeline
opencv-python>=4.8.0
numpy>=1.25.0
pytesseract>=0.3.10
requests>=2.31.0

# Capture
dxcam>=0.0.5

# OCR upgrade (Task 6) — ~500 MB install; CPU-only paddle is fine for text regions
paddleocr>=2.7.0

# UI overlay
PyQt6>=6.6.0

# Observability (Task 8)
loguru>=0.7.0
sentry-sdk>=1.40.0

# Property tests (Task 9)
hypothesis>=6.100.0
